requirements.txt
pyrogram
tgcrypto
loguru
pyyaml
pydantic
# not support windows
# uvloop